Extracted time releated functions to rx-common shared lib ===========================================================
System Information:
Engine Version: Molecule Ver 1.4.1.20210823
Library Version:  Ver 2.0.0.20210823

Firmware:
Windows Ver 1.0.5.20210823
Linux Ver 1.0.3.20210823

Hosts:
Interactive Console Ver 1.0.4.20210823
Pipe Child Ver 1.1.1.20210823
Headless Host Ver 0.9.1.20210823

Storages:
File System Storage Ver 1.1.2.20210823

Compilers:
MSVC 14.29.30133
GCC 9.3.0

OSes:
Linux Ver:5.11.0-27-generic [x86_64]
Windows Workstation 10.0.19041 [x64]

+++ lib/rx_thread.cpp
@@ -537,7 +537,7 @@ void timer::append_job (timer_job_ptr job, job_thread* executer)
 
 void timer::init_random_sequences ()
 {
-	srand(rx_get_tick_count());
+	srand((unsigned int)rx_get_tick_count());
 	for (auto i = 0; i < RX_OFFSET_TIMES_SIZE; i++)
 	{
 		medium_randoms_[i] = (rx_timer_ticks_t)rx_border_rand(0, rx_medium_time_offset);
